% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlFunctions.R
\name{asymptoticControl}
\alias{asymptoticControl}
\title{Control Function for Asymptotic Method Confidence Intervals}
\usage{
asymptoticControl(trans = c("none", "log", "loglog", "logit"), ...)
}
\arguments{
\item{trans}{Transformation apply to the MLE distribution.}

\item{...}{Currently ignored.}
}
\value{
A list with values
\item{\code{trans}}{}
}
\description{
Specify the transformation to apply to the distribution
of the MLE.
}
