% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlFunctions.R
\name{betaControl}
\alias{betaControl}
\title{Control Function for Beta Method for Confidence Intervals}
\usage{
betaControl(...)
}
\arguments{
\item{...}{Currently ignored}
}
\value{
an empty list
}
\description{
In the future, this may provide arguments
that will permit modifications to the Beta method.
}
