% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controlFunctions.R
\name{dobsonControl}
\alias{dobsonControl}
\title{Control Function for Dobson Method Confidence Intervals}
\usage{
dobsonControl(midp = FALSE, tsmethod = c("central", "minlike", "blaker"))
}
\arguments{
\item{midp}{logical, use mid-p values? Currently only permitted
where \code{tsmethod = "central"}.}

\item{tsmethod}{\code{character} giving two-sided method}
}
\value{
a list with values
\itemize{
\item \code{midp}
\item \code{tsmethod}
}
}
\description{
Provides a list of arguments to pass to
\code{\link[exactci]{poisson.exact}}.
}
\seealso{
\code{\link[exactci]{poisson.exact}}
}
