\name{dsrTest-package}
\docType{package}
\title{\packageTitle{dsrTest}}
\description{\packageDescription{dsrTest}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{dsrTest}
  \packageIndices{dsrTest}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{dsrTest}
  
  Maintainer: \packageMaintainer{dsrTest}
}
\references{
Dobson, AJ, Kuulasmaa, K, Eberle, E and Scherer, J (1991)
'Confidence intervals for weighted sums of Poisson parameters',
\emph{Statistics in Medicine}, \strong{10}: 457--462. \doi{doi:10.1002/sim.4780100317}


Swift, MB (1995)  'Simple confidence intervals for standardized rates based on the approximate bootstrap method', \emph{Statistics in Medicine}, \strong{14}, 1875--1888. \doi{doi:10.1002/sim.4780141704}.

Fay MP & Feuer EJ (1997) 'Confidence intervals for directly standardized rates: a method based on the gamma distribution.' \emph{Statistics in Medicine}. \strong{16}: 791--801. \url{https://doi.org/10.1002/(SICI)1097-0258(19970415)16:7<791::AID-SIM500>3.0.CO;2-\%23}

Tiwari RC, Clegg LX, & Zou Z (2006)  'Efficient interval estimation for age-adjusted cancer rates.' \emph{Statistical Methods in Medical Research} \strong{15}: 547--569. \doi{doi:10.1177/0962280206070621}

Ng HKT, Filardo, G & Zheng G (2008) 'Confidence interval estimating procedures for standardized incidence rates.' \emph{Computational Statistics and Data Analysis} \strong{52}: 3501--3516. \doi{doi:10.1016/j.csda.2007.11.004}

Fay, MP & Kim S (2017) 'Confidence intervals for directly standardized rates using mid-p gamma intervals.' \emph{Biometrical Journal}  \strong{59}(2): 377--387. \doi{doi:10.1002/bimj.201600111}

}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
